+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 2026-07-08
 - Provide link for OIDC logout to user after logout from Rqwatch if OPENIDC_RP_INITIATED_LOGOUT is false
+- Redirect back to Rqwatch after manual OIDC logout
 
 ## Version v1.7.2 - Released: 2026-07-09