+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2026-07-12
+- Support LDAP mail alias from alternative attribute
+
 ### 2026-07-11
 - Log session id in login/logout information
 - Get dynamic aliases from multivalued LDAP mail attr